move fragment to shared directory
