+++ apps/editor.planx.uk/src/pages/Team/components/FlowCard/queries.ts
@@ -1,44 +1,7 @@
 import { useMutation } from "@apollo/client";
 import gql from "graphql-tag";
 import { FlowSummary } from "pages/FlowEditor/lib/store/editor";
-import { GET_FLOWS } from "pages/Team/queries";
-
-// this is copied from the getFlows query in FlowEditor/lib/store/editor.ts - might be worth re-using the fragment there as well?
-export const FLOW_SUMMARY_FIELDS = gql`
-  fragment FlowSummaryFields on flows {
-    id
-    name
-    slug
-    status
-    summary
-    updatedAt: updated_at
-    isListedOnLPS: is_listed_on_lps
-    operations(limit: 1, order_by: { created_at: desc }) {
-      createdAt: created_at
-      actor {
-        firstName: first_name
-        lastName: last_name
-      }
-    }
-    templatedFrom: templated_from
-    isTemplate: is_template
-    template {
-      team {
-        id
-        name
-      }
-    }
-    publishedFlows: published_flows(order_by: { created_at: desc }, limit: 1) {
-      publishedAt: created_at
-      hasSendComponent: has_send_component
-      hasVisiblePayComponent: has_pay_component
-      hasEnabledServiceCharge: service_charge_enabled
-    }
-    pinnedFlows: user_pinned_flows {
-      flowId: flow_id
-    }
-  }
-`;
+import { FLOW_SUMMARY_FIELDS, GET_FLOWS } from "pages/Team/queries";
 
 export const PIN_FLOW = gql`
   ${FLOW_SUMMARY_FIELDS}
@@ -87,12 +50,12 @@ interface UnpinFlowVars {
 }
 
 export const usePinFlow = (variables: PinFlowVars) => {
-  const { teamId } = variables
+  const { teamId } = variables;
 
   return useMutation<PinFlowMutation, PinFlowVars>(PIN_FLOW, {
     variables,
     refetchQueries: [{ query: GET_FLOWS, variables: { teamId } }],
-  })
+  });
 };
 
 export const useUnpinFlow = (variables: UnpinFlowVars) => {
@@ -101,4 +64,5 @@ export const useUnpinFlow = (variables: UnpinFlowVars) => {
   return useMutation<UnpinFlowMutation, UnpinFlowVars>(UNPIN_FLOW, {
     variables,
     refetchQueries: [{ query: GET_FLOWS, variables: { teamId } }],
-  })};
+  });
+};
